pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = "Educational project based on \"Django by Example\" book by Antonio Mele"
authors = ["Lex Trofimov <daslef93@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/daslef/django-blog"
repository = "https://github.com/daslef/django-blog"
packages = [
    { include = "django_blog", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5"
django-environ = "^0.11.2"
django-taggit = "^5.0.1"
markdown = "^3.5.2"
pillow = "^10.2.0"
django-ninja = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.1"
mypy = "^1.8.0"
bandit = "^1.7.7"
docformatter = "^1.7.5"
ruff = "^0.2.0"

[tool.poetry.scripts]
django-blog = "django_blog.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
backend-path = ["src"]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = "tests"

[tool.black]
line_length = 88

[tool.ruff]
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
]
ignore = []
line-length = 90
select = [
    "E",
    "F",
    "W",
    "I001",
]

src = ["src", "tests"]

[mypy]
files = ["src"]
strict_optional = false
